requirements.txt
twilio==9.3.4
SQLAlchemy==2.0.34
python-dotenv==1.0.1
pandas==2.2.2
uvicorn==0.32.1
websockets==14.1
fastapi==0.115.5
python-multipart==0.0.17
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.2.1
msgpack==1.1.0
pydantic-settings==2.6.1
faker==33.1.0
pyodbc==5.2.0
httpx==0.28.0
sentry-sdk==2.19.2
python-magic==0.4.27
redis==5.2.1